+++ src/components/Stages/Levels/LevelFour.tsx
@@ -1,32 +1,41 @@
 import React, { useCallback, useContext, useState } from "react";
 import { GameSettingsContext } from "../../../providers/GameSettingsContext";
 import { RoutingContext } from "../../../providers/RoutingContext";
-import { PlaceHolder, Title } from "../../../stories/headers";
+import { Title } from "../../../stories/headers";
 import { PrimaryButton, SkipButton } from "../../../stories/buttons";
 import { useLockNoPrompt } from "../../../utils/lockNavigation";
-
-import { DefaultMessages, LevelFourMessages } from "../../../Messages";
-import DraggingPuzzle from "../../../stories/puzzles/DraggingPuzzle/DraggingPuzzle";
+import { QuestionForm } from "../../../stories/forms";
+import {
+  QuestionFormMessages,
+  TooltipMessages,
+  DefaultMessages,
+  LevelFourMessages,
+} from "../../../Messages";
 
 export const LevelFour = () => {
   const [unLockNavigation, setUnlockNavigation] = useState<boolean>(true);
+  const [skip, setSkip] = useState(false);
   const { skipMode } = useContext(GameSettingsContext);
   const { routeLevelFour } = useContext(RoutingContext);
 
   useLockNoPrompt(unLockNavigation);
 
   const routeChange = useCallback(() => {
-    routeLevelFour();
+    routeLevelFour(skip);
   }, [unLockNavigation]);
 
   const handleUnlockNavigation = () => {
     setUnlockNavigation(false);
   };
 
+  const handleSkip = () => {
+    setSkip(true);
+    setUnlockNavigation(false);
+  };
+
   return (
     <>
-      <PlaceHolder size="small" />
-      <Title titleSize="medium" label={LevelFourMessages.HINT} />
+      <Title titleSize="small" label={LevelFourMessages.HINT} />
       <PrimaryButton
         onClick={routeChange}
         primary={unLockNavigation}
@@ -35,12 +44,19 @@ export const LevelFour = () => {
         data-testid="continueButton"
       />
 
-      <DraggingPuzzle handleUnlockNavigation={handleUnlockNavigation} />
+      <QuestionForm
+        questionIconSize="small"
+        handleUnlock={handleUnlockNavigation}
+        successMessage={QuestionFormMessages.WOW}
+        firstQuestion={QuestionFormMessages.FIRST_Q_LABEL}
+        firstHint={TooltipMessages.FIRST_Q_HINT}
+        firstPlaceholder={QuestionFormMessages.FIRST_Q_PLACEHOLDER}
+        secondQuestion={QuestionFormMessages.SECOND_Q_LABEL}
+        secondHint={TooltipMessages.SECOND_Q_HINT}
+        secondPlaceholder={QuestionFormMessages.SECOND_Q_PLACEHOLDER}
+      />
       {skipMode && (
-        <SkipButton
-          onClick={handleUnlockNavigation}
-          label={DefaultMessages.SKIP_LEVEL}
-        />
+        <SkipButton onClick={handleSkip} label={DefaultMessages.SKIP_LEVEL} />
       )}
     </>
   );
