+++ src/components/Stages/Levels/index.ts
@@ -1,5 +1,6 @@
 export { Start } from "../../Start/Start";
 export { LevelTwo } from "./LevelTwo";
-export { LevelThree } from "./LevelThree";
+export { LevelFour } from "./LevelFour";
 export { LevelOne } from "./LevelOne";
+export { LevelThree } from "./LevelThree";
 export { default as FinishClassic } from "../../Finish/FinishClassic";
